requirements.txt
streamlit
requests
pandas
altair
sqlalchemy
mysql-connector-python
prometheus-client
